+++ psfdx-metadata/psfdx-metadata.Tests.ps1
@@ -38,3 +38,27 @@ Describe 'Retrieve-SalesforceComponent' {
         }
     }
 }
+
+Describe 'Retrieve-SalesforceField' {
+    InModuleScope 'psfdx-metadata' {
+        BeforeEach { Mock Invoke-Salesforce {} }
+        It 'builds custom field retrieve command' {
+            Retrieve-SalesforceField -ObjectName 'Account' -FieldName 'MyField__c' -TargetOrg 'me'
+            Assert-MockCalled Invoke-Salesforce -Times 1 -ParameterFilter {
+                ($Command -eq 'sf project retrieve start --metadata CustomField:Account.MyField__c --target-org me')
+            }
+        }
+    }
+}
+
+Describe 'Retrieve-SalesforceValidationRule' {
+    InModuleScope 'psfdx-metadata' {
+        BeforeEach { Mock Invoke-Salesforce {} }
+        It 'builds validation rule retrieve command' {
+            Retrieve-SalesforceValidationRule -ObjectName 'Account' -RuleName 'MyRule' -TargetOrg 'me'
+            Assert-MockCalled Invoke-Salesforce -Times 1 -ParameterFilter {
+                ($Command -eq 'sf project retrieve start --metadata ValidationRule:Account.MyRule --target-org me')
+            }
+        }
+    }
+}